+++ plugins/all/all.go
@@ -24,6 +24,7 @@ import (
 	_ "github.com/alibaba/ilogtail/plugins/aggregator/shardhash"
 	_ "github.com/alibaba/ilogtail/plugins/aggregator/skywalking"
 	_ "github.com/alibaba/ilogtail/plugins/flusher/checker"
+	_ "github.com/alibaba/ilogtail/plugins/flusher/clickhouse"
 	_ "github.com/alibaba/ilogtail/plugins/flusher/grpc"
 	_ "github.com/alibaba/ilogtail/plugins/flusher/http"
 	_ "github.com/alibaba/ilogtail/plugins/flusher/kafka"