+++ CHANGELOG.md
@@ -45,3 +45,5 @@ your changes, such as:
 - [public] [both] [fixed] ignore timezone adjustment when system time is used
 - [public] [both] [fixed] ignore timezone adjustment when log parse fails
 - [public] [both] [fixed] fix blocking problem caused by alwaysonline config update
+- [public] [both] [added] add a new flusher\_clickhouse plugin.
+